+++ src/utils/clock.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <cstdlib>
+
+// returns the nanosecond (or most precise) performance counter
+size_t clock_nanotime();
