--- src/variations.hpp
@@ -1,4 +1,0 @@
-
-#pragma once
-
-#include "vars/vars_data.hpp"
